move error type *InvalidListFilter* to the new file *errors.go*
